move getClosestCue to texttrack-utils
